tests/requirements.txt
locustio==0.7.2
requests==2.4.3
